Resolve CVE-2023-26115 (#1005)
* Bump transitive dependency on optionator

Upgrade the transitive dependency optionator coming in through ESLint to
the latest available version in order to resolve a transitive dependency
on the word-wrap package following CVE-2023-26115.

* Ignore CVE-2023-26115 for remaining dependencies on word-wrap

The vulnerable package word-wrap is still present in the (dev)dependency
tree through jsfuzz. In this context the vulnerability isn't a problem
so it is ignore.