Use `--fuzzTime` instead of `timeout`
Update the fuzz.js script to accept the --fuzzTime argument and pass it
on to jsfuzz. Update the reusable-fuzz.yml workflow to pass the
--fuzzTime option instead of using timeout (which should make it more
reliable). Given the difference between the --fuzzTime option and the
timeout program, the duration value has been updated to be a numeric
value representing the number of seconds to run (from a string value
representing a time in specified units).