Upgrade `jsfuzz` (#1004)
* Upgrade from `jsfuzz` v1.0.15 to v1.1.2

Replace `jsfuzz` with the `(at)gitlab-org/jsfuzz` in order to upgrade to
the latest available version.

* Update egress policy, allow gitlab

Add gitlab to most CI egress policies following the switch from
depending on the jsfuzz package in the `npmjs.org` registry to the
jsfuzz package in the GitLab npm registry.

* Use `--fuzzTime` instead of `timeout`

Update the fuzz.js script to accept the --fuzzTime argument and pass it
on to jsfuzz. Update the reusable-fuzz.yml workflow to pass the
--fuzzTime option instead of using timeout (which should make it more
reliable). Given the difference between the --fuzzTime option and the
timeout program, the duration value has been updated to be a numeric
value representing the number of seconds to run (from a string value
representing a time in specified units).