Revert to using `timeout` instead of `--fuzzTime` for fuzzing
`--fuzzTime` does not currently work on Windows. For simplicity, just
reverting back to using the `timeout` utility. To make it easier to
switch back in the future the duration in seconds is kept.

This time, just to be sure, update workflows that use the reusable
workflow to use this branch temporarily in order to test the change.