Extract option parsing into a module (#1011)
Refactor the source code to move the options parsing logic out of
index.js and into a new module. This allows for unit test coverage of
the logic (and correspondingly the removal of Stryker disable directive)
allowing for more direct control and testing of that aspect of the
library.

As an aside, this would also allow for reuse of the logic e.g. in the
`shescape/testing` export for more a realistic testing stub.