Fix double quote escaping for CMD with interpolation (#1022)
Update CMD escaping for `{ interpolation: true }` to work similar to
escaping for quoting. This change is motivated by 1) the fact that it is
used when quoting, and 2) that it removes the need for argument
preparation when running e2e tests or fuzzing.

This change fixes incorrect escaping of double quotes for CMD
when interpolation is enabled in certain scenarios.