Separate `node:child_process` options from other options
Define a new option for Shescape that can be used to provide the options
that will be provided to a `node:child_process` instead of spreading
these in the Shescape constructor (and potentially unintentionally
overriding something). This is meant as a defense measure, avoiding
potentially dangerous usage.

Note: this commit is limited to documentation (and types) only.