Update how `child_process` options are given to Shescape
Update documentation as well as test code that relates to taking options
intended for the `child_process` module and providing it to Shescape to
move away from using the spread operator to assigning exactly and only
what needs to be assigned.

The motivation for this is the potential danger of overriding options
unintentionally when using the spread operator. Not using the spread
operator removes this risk entirely. The main benefit of this from this
library's side is simplification of the recipes, which no longer need to
instruct/warn users to be careful when spreading the `child_process`
options into the Shescape options.

The primary motivation for updating testing code accordingly is to
"practice what we preach".