Add proper support for generating a fuzz coverage report
Add a npm command to generate a fuzz coverage command so that someone
fuzzing this project can generate a coverage report to identify missed
branches or other weak points of the fuzz targets.

The command has been added primarily for convenience and for
standardization (e.g. for the .gitignore). But also because the required
`--extension .cjs` option is not obvious (it's necessary because fuzzing
uses the transpiled `index.cjs` module, which nyc won't show by default)