Improve fuzzing experience (#1039)
* Add proper support for generating a fuzz coverage report

Add a npm command to generate a fuzz coverage command so that someone
fuzzing this project can generate a coverage report to identify missed
branches or other weak points of the fuzz targets.

The command has been added primarily for convenience and for
standardization (e.g. for the .gitignore). But also because the required
`--extension .cjs` option is not obvious (it's necessary because fuzzing
uses the transpiled `index.cjs` module, which nyc won't show by default)

* Include --fuzzTime argument in the fuzz script help text

* Add nyc directly

Add nyc[1] as a direct dependency as it is now used for generating
fuzz coverage reports. Previously, the `fuzz:coverage` command worked
because jsfuzz has nyc has a transitive dependency. Adding it as a
direct dependency ensures no unexpected regressions and also makes it
more explicit what this project _actually_ needs to be worked on.

I did try to use (the already present dependency) c8[2] to generate
the fuzz coverage report but this did not seem to work. If you're
seeing this and think it's doable, feel free to submit a patch
removing nyc and updating the `fuzz:coverage` script to use c8.

--
1. https://www.npmjs.com/package/nyc
2. https://www.npmjs.com/package/c8