Make CI fuzzing more fine grained
Set up CI workflows that trigger based on the shell for which the
implementation changed, rather than just the system.

Additionally, this folds the scheduled & manual invocation of CI fuzzing
into the workflows themselves.

Correspondingly, adjust fuzz duration from 5 to 10 min for commit
triggers and from 20 to 10 for scheduled/manual triggers. The former is
considered reasonable because these workflow should trigger less often.
The latter is considered reasonable because longer fuzzing will already
be performed when relevant changes are being made.