Replace depcheck and unimported by knip
Add knip[1] as a devDependency replacing depcheck[2] and unimported[3].
knip provides the same functionality as those two tools combined, and a
bit more. The configuration for knip is largely based on that of
depcheck and unimported. Notable differences are:
- Ignore benches as they're really entry points and also not as
  important (since they change infrequently).
- Ignore binaries lists everything from the `.tool-versions` file (this
  wasn't covered by depcheck nor unimported).
-

The licensee configuration has been updated with 3 new exceptions due
to dependencies introduced in the knip dependency tree:
- deep-freeze: terms described as "public domain" [^1].
- filter-iterator: MIT license, see "License" section of the readme.
- identity-function: terms described as "public domain" [^1].
The existing exception has been removed as it was due to a transitive
dependency of depcheck.

As of this commit, the knip configuration file isn't linted as JSON yet.
This is considered future work, e.g. by switching from `eslint-plugin-
json`[4] to `eslint-plugin-jsonc`[5]/

--
1. https://github.com/webpro/knip
2. https://github.com/depcheck/depcheck
3. https://www.npmjs.com/package/unimported
4. https://www.npmjs.com/package/eslint-plugin-json
5. https://www.npmjs.com/package/eslint-plugin-jsonc