Replace `eslint-plugin-json` by `eslint-plugin-jsonc`
Remove the dependency on `eslint-plugin-json`[1] in favor of `eslint-
plugin-jsonc`[2]. There's a couple of motivations for this, most notably
1) it is more up-to-date (the former's last release was 2 years ago, the
  latter's 1 month), and
2) it has support for `.jsonc` and `.json5`.

The update to the ESLint configuration for the new plugin has been fully
included, the new plugin is completely and explicitly configured. In
general, all rules are enabled and enforce existing styles. A couple of
rules are disabled as they're effectively enforced by Prettier but can't
be configured in a compatible way (this includes `array-bracket-newline`
`array-bracket-spacing`, `array-element-newline`, `object-curly-newline`
`object-curly-spacing`, and `object-property-newline`). The `auto` rule
is omitted because we want to configure everything explicitly. The `key-
name-casing` rule is disabled because I don't think it makes sense for
this use case (linting config files, whose keys we don't control). The
`no-comments` rule is disabled because we'd want to use comments in
JSONC and JSON5 files. Both `sort-array-values` and `sort-keys` are
disabled because alphabetical ordering is not always desired for the
use case of config files.

Lastly, the base ESLint rule `no-irregular-whitespace` has been disabled
for JSON files in accordance with [3], which states that it should be
disabled in favor of the jsonc rule with the same name.

--
1. https://www.npmjs.com/package/eslint-plugin-json
2. https://www.npmjs.com/package/eslint-plugin-jsonc
3. https://ota-meshi.github.io/eslint-plugin-jsonc/rules/no-irregular-whitespace.html