Separate fuzz coverage by target and shell, gain overall coverage data
Update the fuzz script to, on completion, rename the generated coverage
file so that it's unique to the target and shell that was fuzzed. This
results in maintaining coverage data from previous runs. In turn, using
nyc to generate a HTML coverage report will combine all coverage files
into one report, giving a more holistic view of fuzz coverage.

Note: with the current implementation this only really works when jsfuzz
terminates by itself, e.g. due to the usage of `--fuzzTime` or because
of a crash.