Automatically generate the fuzz coverage report
Update the fuzz script to, in addition to organizing the coverage files,
immediately generate the coverage report. Given that this may not happen
correctly, the `npm run fuzz:coverage` command is kept (and used for
this purpose).

The Contributing Guidelines have been updated in accordance with this
changes.

It was considered to update the `reusable-fuzz.yml` workflow to include
the coverage report in the artifact uploaded in the event of a crash,
however the use of `timeout Xs` prevents the `SIGINT` signal hook from
running (so no fuzz coverage report is generated).