Improve use of fuzz coverage data (#1058)
* Separate fuzz coverage by target and shell, gain overall coverage data

Update the fuzz script to, on completion, rename the generated coverage
file so that it's unique to the target and shell that was fuzzed. This
results in maintaining coverage data from previous runs. In turn, using
nyc to generate a HTML coverage report will combine all coverage files
into one report, giving a more holistic view of fuzz coverage.

Note: with the current implementation this only really works when jsfuzz
terminates by itself, e.g. due to the usage of `--fuzzTime` or because
of a crash.

* Send `SIGINT` to jsfuzz when `SIGINT` is received by fuzz.js

Capture the kill (`SIGINT`) signal received by the fuzz.js script and
send it to jsfuzz in order to
1. make sure the pre-exit code (currently moving coverage around) is
   run, and
2. have cleaner exits when the fuzz script is terminated early. I.e. it
   should now be less likely to get a crash when ending a fuzzing
   session manually.

* Automatically generate the fuzz coverage report

Update the fuzz script to, in addition to organizing the coverage files,
immediately generate the coverage report. Given that this may not happen
correctly, the `npm run fuzz:coverage` command is kept (and used for
this purpose).

The Contributing Guidelines have been updated in accordance with this
changes.

It was considered to update the `reusable-fuzz.yml` workflow to include
the coverage report in the artifact uploaded in the event of a crash,
however the use of `timeout Xs` prevents the `SIGINT` signal hook from
running (so no fuzz coverage report is generated).