Improve fuzz time support (#1063)
* Add support for `FUZZ_TIME` environment variable

Update the `fuzz.js` script to consider the `FUZZ_TIME` environment
variable as a secondary source for the fuzz duration in seconds. Use the
static `dotenv/config` import to make sure a `FUZZ_TIME` value set in
`.env` will be used.

Also document this fact in the Contributing Guidelines. In fact, this
actually adds the fact that `--fuzzTime` can be configured to begin with
to the Contributing Guidelines.

* Update log line with information about what fuzzing will do