Update Contributing Guidelines section on testing
Split the section into a subsection on "test organization" (i.e. unit,
integration, etc.) and "writing tests". The former largely covers what
was previously there. Notably, some of the content has been removed and
moved over to the new "writing tests" subsection, and the section on
fuzz testing is now a subsection of the e2e testing section (because the
fuzz tests for this project currently fuzz at the e2e level). Besides
these notable changes, the writing in this section was altered in
general in an attempt to both improve existing copy as well as improve
flow and consistency within the section.

The new "writing tests" section aims to provide some clarity on how
project maintainers think about writing tests. To this end it covers
"test types" (e.g. metamorphic tests) and testing strategies (e.g.
parameterized testing). The former aims to clarify some of the more
advanced tests present in this project as well as encourage contributors
to go beyond the usual oracle test. The latter also aims to clarify some
of the more advanced tests but from the "harness angle" instead of the
"assertion angle".