Remove input validation from `resolveExecutable`
Remove input validation of the provided `deps` from the internal
`resolveExecutable` function. This input validation isn't needed because
the function is only used internally and if it isn't provided this
should be caught by a test or simialr, not an unhelpful runtime check
which unnecessarily impacts performance.

The validation was originally added along the rest of the function in
[1] by me. Unfortunately I did not include a reason for it in the commit
message but I recall it was done for some reason related to testing or
mutation testing. Anyway, testing and mutation testing come back clean
after removal so all should be good.

--
1. 83752afc29fdf7082382f05b4631df024ed272d2