Resolve falsy shell options to a `null` shellName
Update options parsing to resolve falsy shell options to a `null`
shellName to make it internally explicit that we're escaping for a
shell-less context (assuming correct usage of the library). Update
the platform modules to handle this as an expected `shellName`.

In an ideal world, this case would be handled in a way that it only
supports escaping without interpolation (so not escaping with
interpolation or even quoting). However, that would be a breaking change
and hence this change is externally indistinguishable to the original as
the null shellName falls through to the fallback value of getShellName.
Instead, this change aims to make it easier to switch to the above
behavior in the next major version.

As an added benefit it improves performance when the `shell` option is
falsy because it foregoes the executable lookup in favor of a truthy
check.