Resolve falsy shell options to a null shellName
Update options parsing to resolve falsy shell options to a `null`
shellName to make it internally explicit that we're escaping for a
shell-less context (assuming correct usage of the library). Update
the platform modules to handle this as an expected `shellName`. Further
refactoring (see TODOs) is left as for a later commit.

As an added benefit it improves performance when the `shell` option is
falsy because it foregoes the executable lookup in favor of a truthy
check.