Make Shescape reject unsupported shells (#1067)
Change the Shescape implementation such that it now rejects any shells
it does not support, as well as explicitly handle a "no shell" option,
replacing the `interpolation` option. I.e., if the `shell` option is
`false`[^1] interpolation is off and if `shell` is `true` or a `string`
(or an "invalid" value). If `shell` is explicitly configured to an
unsupported shell or `true` and the default system shell isn't supported
then an error will be thrown by the Shescape constructor.

The "no shell" case - `shell: false` - is handled by new shell modules
for both Unix and Windows platforms that are for the no shell or shell-
less use case. These modules don't support quoting (because it isn't
needed and also isn't possible to implement), hence the quote functions
of the Shescape API may now throw for two reasons. Other than that, 
these modules are fairly standard.

All tests, benches, and documentation has (hopefully) been updated
accordingly.

--
^1: deliberately not `falsy` to prevent misuse when `shell` is not
    configured.