Adjust wording for usage of `escape` with `exec`
With v2, shell must always be truthy (actually, always `shell: true` is
even incorrect in general, because it should be the child_process
`shell` option value if present) so the removed statement is a bit
redundant. Instead, more direct language is introduced to indicate that
doing this should really be a last resort.

Also, move up the caveats to increase the likelihood of them being
noticed by the reader.