Require 100 unit test coverage within the project
Configure the `coverage:unit` command to fail if there's not 100%
coverage (function, line, branch, and statement). This helps keep these
test up without the need to rely on an external service like Codecov.

Yes, 100% unit test code coverage is required for this project. For one,
the fact that it's currently achieved shows it's possible. Two, the
code and tests is structured such that 100% coverage is possible. And
three, I believe anything less will result in a sawtooth coverage graph.