Provide explicit $PATH value to which
Update the implementation of `resolveExecutable` to accept the
environment variables so that they can explicitly be provided to
which [1]. All internal code and tests have been updated accordingly,
no external changes.

This is in an effort to fix a problem (bug?) where environment variables
aren't always passed on correctly to subprocesses. The problem is
described in [2], though if Shescape is used in a forked process on
Windows it would face the same problem.

Funny to see that `getShellName` name unit tests for both Unix and
Windows already generated and provided environment variables (:

--
1. https://www.npmjs.com/package/which
2. https://github.com/avajs/ava/issues/3014