Explicitly configure what to `--include` in coverage reports
Update the helper `_coverage` command to explicitly specify what files
should be reported on for coverage. This follows observed coverage
reporting for certain files that aren't of interest.

Notably, this excludes script/, test/, and config files from coverage
reports. The script/ directory may be included because of transpile
(which runs `script/create-d-cts.js`). The test/ directory may be
included because of test helper files (those starting with `_`). Config
files may be included because of tooling used to or before test are run
(e.g. when transpiling the rollup configuration is read).