Convert the e2e (and fuzz) test runners from CJS to ESM
Update test (and related) code to convert the "runners" for e2e (and by
extension fuzz) tests from CJS to ESM. This allows for e2e coverage data
on the ESM source code (over generated CJS code) at the cost of no
coverage guided fuzzing. It also reduces the amount of CJS code to the
absolute minimum, only having fuzz targets in CJS as a requirement of
jsfuzz.