Disable fuzz coverage reports
With the switch to end-to-end testing ESM code, fuzz coverage is no
longer available (coverage is not available across ESM and CJS). Thus,
this removes the use of coverage from fuzzing. This also allows for
dropping the direct dependency on nyc.

This also drops `testing.js` from e2e coverage as it's not e2e tested.