Convert the e2e (and fuzz) test runners from CJS to ESM (#1158)
* Convert the e2e (and fuzz) test runners from CJS to ESM

Update test (and related) code to convert the "runners" for e2e (and by
extension fuzz) tests from CJS to ESM. This allows for e2e coverage data
on the ESM source code (over generated CJS code) at the cost of no
coverage guided fuzzing. It also reduces the amount of CJS code to the
absolute minimum, only having fuzz targets in CJS as a requirement of
jsfuzz.

* Disable fuzz coverage reports

With the switch to end-to-end testing ESM code, fuzz coverage is no
longer available (coverage is not available across ESM and CJS). Thus,
this removes the use of coverage from fuzzing. This also allows for
dropping the direct dependency on nyc.

* Upload e2e test coverage to Codecov

Now possible because coverage is over files present in the repository
(rather than generated CJS files).

* Include all `src/` files in Codecov coverage reports