Create nightly job to test against latest dependency versions (#1169)
* Create nightly job to test against latest dependency versions

Add an (experimental) job to the nightly workflow that will run tests
after updating runtime dependencies to the latest available(&compatible)
version using `npm update`. This is in an effort to ensure compatibility
with runtime dependencies is true for the latest versions, while keeping
compatibility with the oldest version.

The job is run against all OSes given that the dependency `which` might
behave differently based on the OS. The job does not run unit tests
because those _should_ be independent of runtime dependencies.

Also re-ordered installation of shells in all CI workflows to be
alphabetical.

* Sudo should be allowed to install csh and Zsh

* Run `npm install` after update to ensure all dependencies are installed