Split audit npm and secrets workflows
Create a dedicated workflow for secret scanning that is triggered
whenever the `reusable-audit.yml` workflow used to be triggered and
transform the `audit.yml` workflow to be triggered the same way but only
audit the released version if the trigger event is `schedule` or
`workflow_dispatch`. Correspondingly, remove the `reusable-audit.yml`
workflow and update the nightly workflow to not itself run audits (as
this is now handled by the dedicated workflows).

The duplication of the audit job logic is accepted as the benefits
outweigh the drawbacks. Also, the conditional within the job was not a
particularly elegant way to avoid the duplication to begin with.