Audit in CI only if `package-lock.json` changes (#1170)
Create a dedicated workflow for secret scanning that is triggered
whenever the `reusable-audit.yml` workflow used to be triggered and
transform the `audit.yml` workflow to be triggered similarly but for
Pull Requests only if the dependencies could actually have changed. A
separate workflow has been created to audit past releases in order to
avoid skipped jobs on Pull Requests (since these don't need to run on
pushes and Pull Requests).

Correspondingly, remove the `reusable-audit.yml`
workflow and update the nightly workflow to not itself run audits (as
this is now handled by the dedicated workflows).

The duplication of the audit job logic is accepted as the benefits
outweigh the drawbacks. Also, the conditional within the job was not a
particularly elegant way to avoid the duplication to begin with.