Start v2.0.0 (#963)
A quite significant rewrite of the library, but fundamentally the same
in terms of functionality. The API has changed only slight (now
requiring initialize before use), default behavior has changed all
around, and overall the implementation is now safer by default.

Key drivers for the changes in this release are:
- Update runtime dependencies.
- Bump the minimum supported Node.js version.
- Safety by default.
- Better performance on re-use.

All changes were driven by these points. The vast majority or changes
are refactoring to accomodate these points OR refactoring as a result of
changes because of these points.

A full list of (user-facing) changes can be found in the `CHANGELOG.md`.

This commit was preceded by three release candidates(/pre-releases) that
did not result in any problems being reported. Still, this changeset
might not be perfect....