Mutation test `testing.js` module (#1204)
Configure Stryker to mutation test the `testing.js` module. This is done
at the integration level because, like `index.js` it's an exposed
module.

In order to support this, the unit tests for injection strings were
lifted to integration tests in order to provide adequate integration
test coverage to guarantee 100% mutation coverage.