Expand & improve the testing module (#1149)
* Add a throwing test stub to the testing module

* Improve Stubscape

Update the Stubscape implementation so that the behavior of `quote` and
`quoteAll` is more similar to the real implementation of Shescape. Tests
are updated accordingly, mostly resulting in cleaner integration tests
(with more coverage) at the minor cost of more specific unit tests (that
actually more accurately reflect expected behavior).

Also update the Stubscape JSDoc to be up-to-date, both with respect to
the changes seen here as well as the changes of v3.0.0.