Create breakage tests for the testing modules
Expand the breakage test suite with a test file for the shescape/testing
module. This suite is a bit interesting because it contains some skipped
tests. These tests are skipped because of bugfixes and an API extension
which both occurred in [1].

The breakage coverage configuration is also updated to include the
testing module in the coverage report.

--
1. 4f03fd897eef0bcf39ffda6d5f4d0840f5f15c4d