Introduce "breakage testing" (#1233)
Add a new type of testing including a new test suite that's aimed solely
at preventing the introduction of breaking changes. To this end the
suite runs both the current version of the library and the latest
released version and compares their behavior.

This new kind of testing is also documented in the contributing
guidelines (in a bit more detail than the introductory paragraph here).

Note that, for the breakage test file of the shescape/testing module
some tests are skipped. This is because of bugfixes and an API extension
which both occurred in [1].

--
1. 4f03fd897eef0bcf39ffda6d5f4d0840f5f15c4d