Improve breakage tests (#1262)
Add tests verifying that the `Shescape` and `Stubscape` constructors by
themselves behave the same way between the development head and last
released version.

This should improve confidence as without these tests there is a chance
that the implementation is changed such that, e.g., on the development
head the constructor throws where in the last version all methods would
throw instead (which is a breaking change). By having a test dedicated
to the constructor, the described scenario would be caught.

Also some reformatting because I didn't like the way the tests were
written before.