Harden against polluted properties internally (#1285)
Update the internals of the project to protect against the potential of
polluted properties based on manual evaluation of the source code. Re-
use the `hasOwn` function from `options.js` (moved to `reflection.js`)
to achieve this.

- `executables.js`: Explicitly check that `PATH` (or `Path`) isn't being
  inherited. This is the most obvious place where this is necessary as
  the code already accounts for the potential of `PATH` being undefined.
- `platforms.js`: Safely get `OSTYPE` in case it isn't defined in order
  to avoid wrongly concluding the current system is a Windows system as
  a result of a polluted `OSTYPE` value.
- `win.js`: Safely get `ComSpec` in case it isn't defined (already taken
  into account as well) to avoid using a default shell defined by a
  polluted property.

Test suites are updated accordingly to test for the pollution of these
exact properties. (Unfortunately this will only catch regressions, not
potential pollution of other properties. Improvements are welcome.) In
addition, changed test suites have been updated to:

- `resolve.test.js`: remove no-longer-necessary restriction of PATH (or
  Path) being non-empty, in addition to improved test names for clarity.
- `get-helpers.test.js`: reduce duplication.

An important note about the new tests is the use of `fc.pre` in certain
places. This is in order to avoid flakiness due to the (highly unlikely)
scenario where the random actual property value matches the random
polluted value. (note also that these tests don't use `testProp`, this
is because `fc.pre()` does not work with `testProp` in v1.1.6).