Re-export as `Stubscape` from shescape/testing
Simplify importing the default stub implementation of Shescape as
`Stubscape` by re-exporting it under that name.

As of writing, I'm not sure whether exporting this stub as `Shescape` or
`Stubscape` is better. I think both have pros and cons. In any case, the
export named `Shescape` can't just be removed because that would be a
breaking change.