Mark `Stubscape` as an alias of `Shescape`
Avoid vetting error by knip by marking `Stubscape` as an alias of
`Shescape` in its documentation. Besides making knip happy, this also
makes the re-export more explicit and hopefully clear in context.

To support this, `Shescape` was renamed to `Stubscape` and `Shescape`
now re-exports `Shescape`. This makes semantic sense (I want `Stubscape`
to be the "correct" export, and `Shescape` an alias for backwards
compatibility), but is also required as the (at)alias tag needs to be on
the re-exporting expression for knip to accept it.