Re-export as `Stubscape` from shescape/testing (#1308)
Simplify importing the default stub implementation of `Shescape` as
`Stubscape` by re-exporting it under that name.

As of writing, I'm not sure whether exporting this stub as `Shescape` or
`Stubscape` is better. I think both have pros and cons. In any case, the
export named `Shescape` can't just be removed because that would be a
breaking change - so this decision is postponed to the next major
version for now.

Per warnings by knip, `Shescape` is marked as an alias of `Stubscape` in
its JSDoc documentation. Additionally, this aligns the documentation for
`Stubscape` (and by extension `Shescape`) between the .d.ts file and the
.js file.