Update supported versions data
Change the "Supported Versions" section of the Security Policy to add a
end-of-life date for the latest major version. The idea is that this is
extended upon a release so as to 1) set reasonable support expectations
for any individual version, while 2) managing expectations w.r.t. how
long a major version is supported.