Make compatibility tests compatible with Node.js v14.18.0 and v16.13.0
Rewrite the compatibility test suite with a minimalistic home grow test
runner. This provides certainty that it will work with the oldest
supported Node.js version going forward. Since the tests are simple and
don't change often, this trade-off (not using a proper test framework)
is considered worth it.

In the long run, per the `test/compat/runner.js` JSDoc, this may/should
be replaced with the Node.js builtin testing framework. However, that
was not yet available in Node.js v14.18.0 or v16.13.0 [1].

--
1. https://nodejs.org/api/test.html#test-runner