Provide a stateless API (#1130)
Add a separate module which provides a stateless version of the shescape
API. It exposes the same API as an instance of the `Shescape` class,
allowing usage without instantiating a class. This is mostly useful for
use cases where escaping only needs to be done once, e.g. CLI tools.