Fix type resolution for `node10` option
Adjust the export map and transpilation (and correspondingly ignores and
cleaning) to support TypeScript's `node10` type resolution. This
requires that type files for subpath exports (e.g. `shescape/testing`)
are at that exact path (since it doesn't consider the `"exports"` field
of `package.json`). Hence, the build is changed to put these files in
the root of the project, matching the subpath exports.

--
1. https://github.com/arethetypeswrong/arethetypeswrong.github.io/blob/f4cf1648cf8b232fc4e950c2285522e5b57b2108/docs/problems/NoResolution.md#true-positive-node-10-doesnt-support-packagejson-exports