Collect exported modules in `src/modules` (#1362)
Put all the exported modules into `src/modules` to have them grouped and
clean up the root of the project. Update project configuration to
account for this change, notably adjusting ignore files, adjusting paths
for analysis tooling, updating coverage paths, and changing paths in
dev documentation.

Notably, the TypeScript type decleration files, while having been moved,
are programatically moved to the root of the project for releases. This
is to support TypeScript's `node10` type resolution, which requires that
type files for subpath exports (e.g. `shescape/testing`) are at that
exact path (since it doesn't consider the `"exports"` field of
`package.json`). Hence, the build is changed to put these files in the
root of the project, matching the subpath exports. See the following for
more information:
<https://github.com/arethetypeswrong/arethetypeswrong.github.io/blob/f4cf1648cf8b232fc4e950c2285522e5b57b2108/docs/problems/NoResolution.md#true-positive-node-10-doesnt-support-packagejson-exports>