Improve manual Release Guidelines
Update the manual Release Guidelines to include commands/options to
avoid errors while executing the commands while following the steps.

When syncing the repository, dependencies must be installed because the
helper scripts to bump the JSDoc and changelog need them. Similarly,
ensure dependencies are installed when publishing because that's needed
to be able to transpile. This is included separately because of the
delay in between initiating the release and finally publishing (that is,
the opening, and waiting for validation, of the release Pull Request).