Redo alerts (i.e. notes and warnings) in the documentation (#1405)
Update the "alerts" present in the MarkDown documentation to move adopt
a style that works independently of where it is rendered.

The previous format was an old variant of the GitHub flavored markdown
syntax that created properly rendered alerts on GitHub. This syntax has
changed since, however the new one is
1. Not as clear outside the context of GitHub, and
2. Not implemented correctly (it doesn't render correctly in all
   scenarios, e.g. it wouldn't render properly for the use case in the
   Release Guidelines).

Given those considerations, this adopts a semantically correct (no block
quotes) and standard-markdown way to have alerts (i.e. text blocks that
stand out from the rest of the text in some way).