Integration test for "shell missing" and "shell unsupported" (#1407)
Shell should be a program that exists on any system that runs this test
suite but not a shell that is actually supported. The Node.js binary
should exist when this test suite is run, but should not be a shell that
is supported.

This adds coverage for the `if (!isShellSupported(shellName))` branch in
`src/internal/options.js` (and correspondingly the `unsupportedError`
function).