Add "shell missing" and "shell unsupported" cases to arbitrary options
Update the `shescapeOptions` arbitrary to be able to generate options
with either a missing shell `"not-actually-a-shell-that-exists"` or a
shell that (should) exists but isn't supported `"node"`.

Note that the Node.js binary should exist when test are run, but should
not be a shell that is supported by Shescape.

This ensures increased coverage in `src/internal/options.js` and
`src/internal/executables.js`.

This is a generalized version of the changes made in
d4c0438bb1e94332f270a883eef1a4f29034645b