Resolve helper script issues on Windows starting wtih Node.js v20.12.2 (#1515)
This fixes an issue with this project's helper scripts that started in
Node.js v20.12.2 due to the fix for CVE-2024-27980. In particular, we
spawn using `npm.cmd` on Windows to be able to run npm scripts
programmatically (just `npm` doesn't work on Windows), however because
of the CVE this is no longer allowed without setting `shell: true`. In
this context, setting `shell: true` doesn't really matter (all input is
considered trusted) so we just do that.