Move to fuzzing with fast-check
Remove jsfuzz and use fast-check for fuzzing instead. There are several
advantages to this given the state of the jsfuzz project (i.e. it's not
well-maintained). The drawbacks are limited since the fuzzing for this
project is largely random(/unguided) anyway. The advantage of fast-check
is that 1) it does not introduce a new dependency and 2) like jsfuzz it
does not need any system dependency (in contract to Jazzer.js), and 3)
it support ESM modules (finally getting rid of the last remaining CJS
source code from this project).

One major difference between fuzzing with jsfuzz (or Jazzer.js) and
fast-check is that test duration is better specified in terms of test
runs/iterations rather than duration. This actually seems fine (maybe
even better) to me.