Pin Node.js version for development
Update `.nvmrc` to pin the Node.js version to be used during
development. Obviously anyone is still free to use any version in the
same major version range (though earlier versions might not work). The
motivation for this is two incidents where the latest Node.js release
had problems within the span of one year (currently, Node.js 22.5.0 is
latest and has a breaking bug when installing dependencies, in Sept of
last year 20.6.0 had a bug affecting development tooling [1]).

This overrides the change made in [2] which moved from a pinned version
to a major version range. The commit is light on details but I believe I
made this decision because I didn't want to force anyone to use a
specific version, as long as it's 20 (then, 22 now) its fine. I still
believe this, but unfortunately being on the edge all the time is a bit
too unstable.

I don't see any particular need to stay on the edge, at least for this
project which is receiving fairly few updates. Hence, I do not intent to
regularly update the version used - probably not until the next major
LTS release.

--
1. 9c96c652fa3cc80afdb4d8861860f46879539935
2. 02d4f1cc7d9cebad3938d21bb0b73630f18b08a4